hid panel states, no longer a true representation of requirments or needed objects
added a perminant add/create button for themap panel
this is a special case because the model can only have one map but may need to edit and create a new one at the same time.

added edit functionality for the map, specifically parameterized currently, doens't seem to be working during an edit server is returning a 415, will notify jay.