Added 3 second pause before starting fate model run
This is to avoid a race condition that occurs when the page is refreshed
while a model is currently running. The model rewind and greenlet kill
do not happen before a new connection and model start are requested. By
pausing for 3 seconds after page load before a new model run is
requested, this gives time for the rewind to occur.