Changing "Name" value in edit wind form (form/wind.js) now updates both mover name and enviro obj name.
Note: this now makes it impossible to change the enviro obj name in the "advanced settings" since it gets over written when the form is rendered (the atts are updated from the form) - but its better if enviro obj name matches mover name anyway. May want to revisit this in future if we need more separation.

Other minor fixes:
renamed title in current edit form
bugfix: scale_factor was renamed current_scale in pygnome